get necessary data for pages
