+++ src/cloud/pages/automations/[automationId].tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import { getAutomation } from '../../api/automation/automation'
+import { getTeamIndexPageData } from '../../api/pages/teams'
 import { SerializedAutomation } from '../../interfaces/db/automations'
 import { GetInitialPropsParameters } from '../../interfaces/pages'
 
@@ -15,12 +16,15 @@ const AutomationPage = ({
   )
 }
 
-AutomationPage.getInitialProps = async ({
-  pathname,
-}: GetInitialPropsParameters) => {
-  const workflowId = pathname.split('/')[3]
+AutomationPage.getInitialProps = async (params: GetInitialPropsParameters) => {
+  const workflowId = params.pathname.split('/')[3]
+  const [teamData, automation] = await Promise.all([
+    getTeamIndexPageData(params),
+    getAutomation(workflowId),
+  ])
   return {
-    workflow: await getAutomation(workflowId),
+    ...teamData,
+    automation,
   }
 }
