+++ src/cloud/pages/automations/create.tsx
@@ -1,7 +1,10 @@
 import React from 'react'
+import { getTeamIndexPageData } from '../../api/pages/teams'
 
 const AutomationCreatePage = () => {
   return <div>Create</div>
 }
+
+AutomationCreatePage.getInitialProps = getTeamIndexPageData
 
 export default AutomationCreatePage
