+++ src/cloud/pages/workflows/[workflowId].tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import { getWorkflow } from '../../api/automation/workflow'
+import { getTeamIndexPageData } from '../../api/pages/teams'
 import { SerializedWorkflow } from '../../interfaces/db/automations'
 import { GetInitialPropsParameters } from '../../interfaces/pages'
 
@@ -11,12 +12,15 @@ const WorkflowPage = ({ workflow }: { workflow: SerializedWorkflow }) => {
   )
 }
 
-WorkflowPage.getInitialProps = async ({
-  pathname,
-}: GetInitialPropsParameters) => {
-  const workflowId = pathname.split('/')[3]
+WorkflowPage.getInitialProps = async (params: GetInitialPropsParameters) => {
+  const workflowId = params.pathname.split('/')[3]
+  const [teamData, workflow] = await Promise.all([
+    getTeamIndexPageData(params),
+    getWorkflow(workflowId),
+  ])
   return {
-    workflow: await getWorkflow(workflowId),
+    ...teamData,
+    workflow,
   }
 }
