+++ src/cloud/pages/workflows/create.tsx
@@ -1,7 +1,10 @@
 import React from 'react'
+import { getTeamIndexPageData } from '../../api/pages/teams'
 
 const WorkflowCreatePage = () => {
   return <div>Create</div>
 }
+
+WorkflowCreatePage.getInitialProps = getTeamIndexPageData
 
 export default WorkflowCreatePage
